cardodge refactor, cardodge settings integration, game command, help command update
