+++ GameFiles/CarDodge/CarDodge_CodeFiles/CarInfo/CarInfo.cpp
@@ -1,0 +1,5 @@
+//
+// CarInfo.cpp - Houses the CarInfo and CarStyles structure functions.
+//
+
+#include "CarInfo.h"
